Peaks.jl

Peaks.argmaxima — Function

argmaxima(x[, w=1; strict=true])

Find the indices of the local maxima of x where each maxima i is either the maximum of x[i-w:i+w] or the first index of a plateau.

If strict is true, all elements of x[i-w:i+w] or the bounds of a plateau must exist and must not be missing or NaN. For strict == false, a maxima is the maximum or first element of all existing and non-NaN or missing elements in x[i-w:i+w] or the bounds of a plateau.

See also: findmaxima, findnextmaxima

Examples

julia> argmaxima([0,2,0,1,1,0])
2-element Vector{Int64}:
 2
 4

julia> argmaxima([2,0,1,1])
Int64[]

julia> argmaxima([2,0,1,1]; strict=false)
2-element Vector{Int64}:
 1
 3

Peaks.argminima — Function

argminima(x[, w=1; strict=false])

Find the indices of the local minima of x where each minima i is either the minimum of x[i-w:i+w] or the first index of a plateau.

If strict is true, all elements of x[i-w:i+w] or the bounds of a plateau must exist and must not be missing or NaN. For strict == false, a minima is the minimum or first element of all existing and non-NaN or missing elements in x[i-w:i+w] or the bounds of a plateau.

See also: findminima, findnextminima

Examples

julia> argminima([3,2,3,1,1,3])
2-element Vector{Int64}:
 2
 4

julia> argminima([2,3,1,1])
Int64[]

julia> argminima([2,3,1,1]; strict=false)
2-element Vector{Int64}:
 1
 3

Peaks.findmaxima — Function

findmaxima(x[, w=1; strict=true]) -> (idxs, vals)

Find the indices and values of the local maxima of x where each maxima i is either the maximum of x[i-w:i+w] or the first index of a plateau.

See also: argmaxima, findnextmaxima

Peaks.findminima — Function

findminima(x[, w=1; strict=true]) -> (idxs, vals)

Find the indices and values of the local minima of x where each minima i is either the minimum of x[i-w:i+w] or the first index of a plateau.

See also: argminima, findnextminima

Missing docstring.

Missing docstring for peakprom. Check Documenter's build log for details.

Missing docstring.

Missing docstring for peakprom!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for peakwidth. Check Documenter's build log for details.

Missing docstring.

Missing docstring for peakwidth!. Check Documenter's build log for details.

Peaks.findnextmaxima — Function

findnextmaxima(x, i[, w=1; strict=true])

Find the index of the next maxima in x after or including i, where the maxima i is either the maximum of x[i-w:i+w] or the first index of a plateau. Returns lastindex(x) + 1 if no maxima occur after i.

If strict is true, all elements in x[i-w:i+w] or the bounds of a plateau must exist and must not be missing or NaN. For strict == false, a maxima is the maximum or first element of all existing and non-NaN or missing elements in x[i-w:i+w] or the bounds of a plateau.

See also: argmaxima

Examples

julia> findnextmaxima([0,2,0,1,1,0], 2)
2

julia> findnextmaxima([0,2,0,1,1,0], 3)
4

Peaks.findnextminima — Function

findnextminima(x, i[, w=1, strict=true])

Find the index of the next minima in x after or including i, where the minima i is either the minimum of x[i-w:i+w] or the first index of a plateau. Returns lastindex(x) + 1 if no minima occur after i.

If strict is true, all elements in x[i-w:i+w] or the bounds of a plateau must exist and must not be missing or NaN. For strict == false, a minima is the minimum or first element of all existing and non-NaN or missing elements in x[i-w:i+w] or the bounds of a plateau.

See also: argminima

Examples

julia> findnextminima([3,2,3,1,1,3], 2)
2

julia> findnextminima([3,2,3,1,1,3], 3)
4